+++ src/demo/Demo.vue
@@ -8,7 +8,7 @@ import { demos } from './utils/mocks.js'
   <TvDemo
     :component="TvSidebar"
     :variants="demos"
-    :manual-emits="['clickLabel', 'click', 'clickImage']"
+    :manual-emits="['click']"
     hide-background
     component-name="TvSidebar"
     npm-install="@todovue/tv-sidebar"